+++ docs/source/tutorial/warning_ctrl/filter_warnings_demo.py
@@ -1,0 +1,12 @@
+import warnings
+from deprecated import deprecated
+
+
+@deprecated(version='1.2.1', reason="deprecated function")
+def fun():
+    print("fun")
+
+
+if __name__ == '__main__':
+    warnings.simplefilter("ignore", category=DeprecationWarning)
+    fun()